Moved Bee Stinger and Web Spinner back to Tail, where they probably belong. introduced ovipositor as new body type. Added knockup event which can be subscribed to, useful primarily for heat status effect. added heat, rut, slime craving, status effects, as well as elastic innards and vag of holding. Hyper happy setting introduced. Transformations introduced, with generic means to allow anything to proc them.
introduced diapause option to all wombs and all pregnancy stores. all pregnancies will now take into account diapause, if applicable.

Fixed a bug introduced in previous update that did not update the text correctly for option row buttons where the description would not update, even though the actual value did.

+++ CoC.Backend/Creatures/CombatCreature.cs
@@ -615,7 +615,7 @@ namespace CoC.Backend.Creatures
 			}
 		}
 
-		public void DeltaCombatCreatureStats(short lus = 0, short lib = 0, short sens = 0, short corr = 0, short str = 0, short tou = 0, short spe = 0, short inte = 0, bool ignorePerks = false)
+		public void DeltaCombatCreatureStats(float lus = 0, float lib = 0, float sens = 0, float corr = 0, float str = 0, float tou = 0, float spe = 0, float inte = 0, bool ignorePerks = false)
 		{
 			DeltaCreatureStats(lust, lib, sens, corr);
